Divided tests into separate files
