+++ tests/retrieve_results_test.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+
+import unittest
+
+import venpy
+
+class TestRetrieveResults(unittest.TestCase):
+    
+    def test_expected_result(self):
+        pass
+    
+    def test_names_passed_result(self):
+        pass
+    
+    def test_vtype_passed_result(self):
+        pass
+    
+    def test_both_names_vtype_passed_result(self):
+        pass
